+++ src/IReadonlyRepository.ts
@@ -1,6 +1,7 @@
 import type { Entity } from './Entity';
 import type { ModelMetadata } from './metadata';
 import type { CountResult, FindArgs, FindOneArgs, FindOneResult, FindResult, WhereQuery } from './query';
+import type { CountArgs } from './query/CountArgs';
 import type { QueryResult } from './types';
 
 export interface IReadonlyRepository<T extends Entity> {
@@ -12,6 +13,7 @@ export interface IReadonlyRepository<T extends Entity> {
    * @param {string[]} [args.select] - Array of model property names to return from the query.
    * @param {object} [args.where] - Object representing the where query
    * @param {string|object|string[]|object[]} [args.sort] - Property name(s) to sort by
+   * @param {object} [args.pool] - Override the db pool to use for the query
    */
   findOne<
     // Optional keys specified as args.select
@@ -30,6 +32,7 @@ export interface IReadonlyRepository<T extends Entity> {
    * @param {string|object|string[]|object[]} [args.sort] - Property name(s) to sort by
    * @param {string|number} [args.skip] - Number of records to skip
    * @param {string|number} [args.limit] - Number of results to return
+   * @param {object} [args.pool] - Override the db pool to use for the query
    */
   find<
     // Optional keys specified as args.select
@@ -42,8 +45,10 @@ export interface IReadonlyRepository<T extends Entity> {
 
   /**
    * Gets a count of rows matching the where query
-   * @param {object} [where] - Object representing the where query
+   * @param {object} [args] - Arguments
+   * @param {object} [args.where] - Object representing the where query
+   * @param {object} [args.pool] - Override the db pool to use for the query
    * @returns {number} Number of records matching the where criteria
    */
-  count(where?: WhereQuery<T>): CountResult<T>;
+  count(args?: CountArgs<T> | WhereQuery<T>): CountResult<T>;
 }